completed excption handling except regex

+++ src/main/java/com/yash/shopkeeper/Shopkeeper/Profile/controller/ShopkeeperController.java
@@ -4,8 +4,7 @@ import com.yash.shopkeeper.Shopkeeper.Profile.dto.ShopkeeperLoginDto;
 import com.yash.shopkeeper.Shopkeeper.Profile.dto.ShopkeeperRegisterDto;
 import com.yash.shopkeeper.Shopkeeper.Profile.dto.ShopkeeperUpdatePasswordDto;
 import com.yash.shopkeeper.Shopkeeper.Profile.entity.ShopkeeperProfile;
-import com.yash.shopkeeper.Shopkeeper.Profile.exceptions.AlreadyPresentShopkeeperException;
-import com.yash.shopkeeper.Shopkeeper.Profile.exceptions.RegisterPayloadException;
+import com.yash.shopkeeper.Shopkeeper.Profile.exceptions.*;
 import com.yash.shopkeeper.Shopkeeper.Profile.service.ShopkeeperService;
 import com.yash.shopkeeper.Shopkeeper.Profile.utilities.PayloadCheck;
 import org.apache.logging.log4j.LogManager;
@@ -31,7 +30,19 @@ public class ShopkeeperController {
 
     public static String REGISTER_PAYLOAD_WRONG = "THE REGISTER PAYLOAD IS NOT CORRECT FOR THE SHOPKEEPER";
 
+    public static String LOGIN_PAYLOAD_WRONG = "THE LOGIN PAYLOAD IS NOT CORRECT FOR THE SHOPKEEPER";
+
+    public static String WRONG_CREDENTIALS_FOR_THE_USER = "The emailId or password is incorrect for the user";
+
     public static String ALREADY_PRESENT_SHOPKEEPER = "This shopkeeper is already present in the DB";
+
+    public static String USERID_PAYLOAD_WRONG = "The payload for getShopkeeperById is not correct";
+
+    public static String SHOPKEEPER_NOT_FOUND = "The shopkeeper is not found for the given userId";
+
+    public static String UPDATE_PASSWORD_PAYLOAD_WRONG = "The update password payload is not correct";
+
+    public static String EMAILID_NOT_FOUND = "The emailId not found for the user";
     @Value("${spring.application.name}")
     String applicationName;
 
@@ -61,8 +72,19 @@ public class ShopkeeperController {
 
     @GetMapping(value = "/shopKeeperById/{userId}", headers = "Accept=application/json")
     public ShopkeeperProfile getShopKeeperById(@PathVariable String userId) {
-        log.info("Getting the shopKeeper profile by the userId");
-        ShopkeeperProfile shopkeeperProfile = shopkeeperService.getShopKeeperById(userId);
+        log.info("Checking for the payload for getShopkeeper by ID API");
+        ShopkeeperProfile shopkeeperProfile = null;
+        if(!payloadCheck.isUserIdValid(userId)){
+            log.info("The userId is not valid for the getShopkeeperById");
+            throw new UserIdNullException(USERID_PAYLOAD_WRONG);
+        }
+        try {
+            log.info("Getting the shopKeeper profile by the userId");
+            shopkeeperProfile = shopkeeperService.getShopKeeperById(userId);
+        }catch (ShopkeeperNotFound shopkeeperNotFound){
+            log.info("The shopkeeper is not found for this.");
+            throw new ShopkeeperNotFound(SHOPKEEPER_NOT_FOUND);
+        }
         log.info("The shopKeeperProfile is here");
         return shopkeeperProfile;
     }
@@ -82,39 +104,57 @@ public class ShopkeeperController {
 
     @PutMapping(value = "/updatePassword", headers = "Accept=application/json")
     public boolean updatePassword(@RequestBody ShopkeeperUpdatePasswordDto shopkeeperUpdatePasswordDto) {
-        log.info("Updating the password");
-        boolean isUpdated = shopkeeperService.updatePassword(shopkeeperUpdatePasswordDto);
-        if (isUpdated) {
-            return true;
-        } else {
-            return false;
+        log.info("Checking for the payload for updating the password");
+        boolean isUpdated = false;
+        if(!payloadCheck.isUpdatePasswordPayloadValid(shopkeeperUpdatePasswordDto)){
+            log.info("The update password payload is not valid");
+            throw new UpdatePasswordPayloadException(UPDATE_PASSWORD_PAYLOAD_WRONG);
         }
+        try {
+            log.info("Updating the password");
+             isUpdated = shopkeeperService.updatePassword(shopkeeperUpdatePasswordDto);
+        }catch (EmailIdNotFoundException e){
+            log.error("The emailId not found for the user");
+            throw new EmailIdNotFoundException(EMAILID_NOT_FOUND);
+        }
+        return true;
     }
 
 
     @DeleteMapping(value = "/deleteShopkeeperProfileByUserId/{userId}", headers = "Accept=application/json")
     public boolean deleteShopkeeperProfileByUserId(@PathVariable String userId) {
-        log.info("Deleting the shopkeeper profile by the userId");
-        boolean isDeleted = shopkeeperService.deleteShopkeeper(userId);
-        if (isDeleted) {
-            log.info("The user is deleted");
-            return true;
-        } else {
-            log.error("The user is not deleted");
-            return false;
+        boolean isDeleted = false;
+        log.info("Checking for the payload of deleteShopkeeperByUserId");
+        if(!payloadCheck.isDeleteShopkeeperByUserIdValid(userId)){
+            log.info("The userId is not valid for the getShopkeeperById");
+            throw new UserIdNullException(USERID_PAYLOAD_WRONG);
         }
+        try {
+            log.info("Deleting the shopkeeper profile by the userId");
+            isDeleted = shopkeeperService.deleteShopkeeper(userId);
+        }catch (ShopkeeperNotFound shopkeeperNotFound){
+            throw new ShopkeeperNotFound("There is no shopkeeper profile by this userId");
+        }
+        return true;
     }
 
     @PostMapping(value = "/login", headers = "Accept=application/json")
     public ShopkeeperProfile loginShopkeeper(@RequestBody ShopkeeperLoginDto shopkeeperLoginDto) {
-        ShopkeeperProfile shopkeeperProfile = shopkeeperService.loginShopkeeper(shopkeeperLoginDto);
-        if (shopkeeperProfile == null) {
-            log.error("Unable to login");
-            return null;
-        } else {
-            log.info("The login is successful");
-            return shopkeeperProfile;
+        log.info("Checking for the payload of login API");
+        ShopkeeperProfile shopkeeperProfile = null;
+        if(!payloadCheck.isLoginPayloadValid(shopkeeperLoginDto)){
+            log.info("The login payload is not valid");
+            throw new LoginPayloadException(LOGIN_PAYLOAD_WRONG);
         }
+        try {
+            log.info("The login payload for the shopkeeper is correct");
+            shopkeeperProfile = shopkeeperService.loginShopkeeper(shopkeeperLoginDto);
+        }catch (WrongCredentialsForShopkeeper wrongCredentialsForShopkeeper){
+            log.info("The emailId or password is incorrect");
+            throw new WrongCredentialsForShopkeeper(WRONG_CREDENTIALS_FOR_THE_USER);
+        }
+        log.info("The login is successful");
+        return shopkeeperProfile;
     }
 
 }
